Rotator angle is a double.

+++ Lumidex.Core/Data/ImageFile.cs
@@ -69,7 +69,7 @@ public class ImageFile
     [Column(TypeName = "TEXT COLLATE NOCASE")]
     public string? RotatorName { get; set; }
 
-    public int? RotatorPosition { get; set; }
+    public double? RotatorPosition { get; set; }
 
     #endregion
